# introduce the concept of deferred stitching