--- a/plugins/catalog-backend/src/stitching/progressTracker.ts
+++ b/plugins/catalog-backend/src/stitching/progressTracker.ts
@@ -19,10 +19,11 @@ import { metrics } from '@opentelemetry/api';
 import { Knex } from 'knex';
 import { DateTime } from 'luxon';
 import { Logger } from 'winston';
+import { DbRefreshStateRow } from '../database/tables';
 import { createCounterMetric } from '../util/metrics';
 
 // Helps wrap the timing and logging behaviors
-export function progressTracker(_knex: Knex, logger: Logger) {
+export function progressTracker(knex: Knex, logger: Logger) {
   // prom-client metrics are deprecated in favour of OpenTelemetry metrics.
   const promStitchedEntities = createCounterMetric({
     name: 'catalog_stitched_entities_count',
@@ -46,6 +47,26 @@ export function progressTracker(_knex: Knex, logger: Logger) {
     },
   );
 
+  const stitchingQueueCount = meter.createObservableGauge(
+    'catalog.stitching.queue.length',
+    { description: 'Number of entities currently in the stitching queue' },
+  );
+  stitchingQueueCount.addCallback(async result => {
+    const total = await knex<DbRefreshStateRow>('refresh_state')
+      .count({ count: '*' })
+      .whereNotNull('next_stitch_at');
+    result.observe(Number(total[0].count));
+  });
+
+  const stitchingQueueDelay = meter.createHistogram(
+    'catalog.stitching.queue.delay',
+    {
+      description:
+        'The amount of delay between being scheduled for stitching, and the start of actually being stitched',
+      unit: 'seconds',
+    },
+  );
+
   function stitchStart(item: {
     entityRef: string;
     stitchRequestedAt?: DateTime;
@@ -53,6 +74,11 @@ export function progressTracker(_knex: Knex, logger: Logger) {
     logger.debug(`Stitching ${item.entityRef}`);
 
     const startTime = process.hrtime();
+    if (item.stitchRequestedAt) {
+      stitchingQueueDelay.record(
+        -item.stitchRequestedAt.diffNow().as('seconds'),
+      );
+    }
 
     function endTime() {
       const delta = process.hrtime(startTime);