--- a/plugins/catalog-backend/src/stitching/types.ts
+++ b/plugins/catalog-backend/src/stitching/types.ts
@@ -14,6 +14,9 @@
  * limitations under the License.
  */
 
+import { Config } from '@backstage/config';
+import { HumanDuration } from '@backstage/types';
+
 /**
  * Performs the act of stitching - to take all of the various outputs from the
  * ingestion process, and stitching them together into the final entity JSON
@@ -33,8 +36,45 @@ export interface Stitcher {
  * @remarks
  *
  * In immediate mode, stitching happens "in-band" (blocking) immediately when
- * each processing task finishes.
+ * each processing task finishes. When set to `'deferred'`, stitching is instead
+ * deferred to happen on a separate asynchronous worker queue just like
+ * processing.
+ *
+ * Deferred stitching should make performance smoother when ingesting large
+ * amounts of entities, and reduce p99 processing times and repeated
+ * over-stitching of hot spot entities when fan-out/fan-in in terms of relations
+ * is very large. It does however also come with some performance cost due to
+ * the queuing with how much wall-clock time some types of task take.
  */
-export type StitchingStrategy = {
-  mode: 'immediate';
-};
+export type StitchingStrategy =
+  | {
+      mode: 'immediate';
+    }
+  | {
+      mode: 'deferred';
+      pollingInterval: HumanDuration;
+      stitchTimeout: HumanDuration;
+    };
+
+export function stitchingStrategyFromConfig(config: Config): StitchingStrategy {
+  const strategyMode = config.getOptionalString(
+    'catalog.stitchingStrategy.mode',
+  );
+
+  if (strategyMode === undefined || strategyMode === 'immediate') {
+    return {
+      mode: 'immediate',
+    };
+  } else if (strategyMode === 'deferred') {
+    // TODO(freben): Make parameters configurable
+    return {
+      mode: 'deferred',
+      pollingInterval: { seconds: 1 },
+      stitchTimeout: { seconds: 60 },
+    };
+  }
+
+  throw new Error(
+    `Invalid stitching strategy mode '${strategyMode}', expected one of 'immediate' or 'deferred'`,
+  );
+}