--- a/plugins/catalog-backend/src/tests/integration.test.ts
+++ b/plugins/catalog-backend/src/tests/integration.test.ts
@@ -215,6 +215,11 @@ class TestHarness {
             connection: ':memory:',
           },
         },
+        catalog: {
+          stitchingStrategy: {
+            mode: 'immediate',
+          },
+        },
       },
     );
     const logger = options?.logger ?? getVoidLogger();