… .changeset/loud-ghosts-deny.md
… plugins/catalog-backend/src/database/operations/stitcher/getDeferredStitchableEntities.test.ts
… plugins/catalog-backend/src/database/operations/stitcher/markForStitching.test.ts
… plugins/catalog-backend/src/database/operations/stitcher/performStitching.test.ts
… plugins/catalog-backend/src/database/operations/util/deleteOrphanedEntities.test.ts
… plugins/catalog-backend/src/tests/migrations.test.ts
… plugins/catalog-backend/src/tests/performance/stitchingPerformance.test.ts