## From the changeset

Introduce a new optional config parameter `catalog.stitchingStrategy.mode`, which can have the values `'immediate'` (default) and `'deferred'`. The default is for stitching to work as it did before this change, which means that it happens "in-band" (blocking) immediately when each processing task finishes. When set to `'deferred'`, stitching is instead deferred to happen on a separate asynchronous worker queue just like processing.

Deferred stitching should make performance smoother when ingesting large amounts of entities, and reduce p99 processing times and repeated over-stitching of hot spot entities when fan-out/fan-in in terms of relations is very large. It does however also come with some performance cost due to the queuing with how much wall-clock time some types of task take.

## Timing results

First: running the migration on a 250 000 entity database with three relations each, took 11 seconds on my local docker based postgres 13 database.

Below are the results of running both immediate and deferred mode with different synthetic large ingestions. Steady state is not expected to be significantly different - the main impact there is that a lag of up to one second is added between processing and stitching.

* **25000 entities, 0 relations each**

  |  | Immediate (original) | Deferred (this PR) |
  | - | - | - | 
  | Total stitcher runs | 25000 | 25000 | 
  | Wall clock time | 1 min 15 sec | 1 min 48 sec | 
  | Speed |  | 44% slower | 

  Few things have no relations whatsoever, but this case probably shows some form of "maximum slowdown" numbers.

* **5000 entities, 3 relations each**

  |  | Immediate (original) | Deferred (this PR) |
  | - | - | - | 
  | Total stitcher runs | 19994 | 7608 | 
  | Wall clock time | 29 sec | 29 sec | 
  | Speed |  | Same | 

  At this scale the wall clock time is the same, but the number of stitch operations is still much lower which potentially puts less load on your database - with the caveat that the queue mechanism in the deferred mode performs db ops too, of course.

* **25000 entities, 2 relations each**

  |  | Immediate (original) | Deferred (this PR) |
  | - | - | - | 
  | Total stitcher runs | 75000 | 37266 | 
  | Wall clock time | 1 min 59 sec | 2 min 13 sec | 
  | Speed |  | 12% slower | 

  Some time penalty when the number of relations is less than three.

* **5000 entities, all with 1 relation to a single "victim" entity**

  |  | Immediate (original) | Deferred (this PR) |
  | - | - | - | 
  | Total stitcher runs | 10001 | 5002 | 
  | Wall clock time | 2 min 8 sec | 25 sec | 
  | Speed |  | over 5x faster |

  This pathological highly connected case shows the original immediate mode code slowing down more and more until it's at one tenth of its original full speed near the completion of the test, because it becomes completely overburdened with repeatedly re-stitching the same thing over and over again, competing for the exact same rows in the database.

* **25000 entities, all with 1 relation to a single "victim" entity**

  |  | Immediate (original) | Deferred (this PR) |
  | - | - | - | 
  | Total stitcher runs | 50001 | 25002 | 
  | Wall clock time | 50 min | 1 min 44 sec | 
  | Speed |  | over 28x faster |

  Clearly the old solution slows down more and more as the connectivity rises, while the new solution is closer to linear.